changes in todo.py

+++ TODO/routers/todos.py
@@ -41,29 +41,36 @@ user_dependency = Annotated[dict,Depends(get_current_user)]
 
 @router.post("/todos", status_code=status.HTTP_201_CREATED)
 def create_todo(user: user_dependency,todo: TodoSchema, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Please login first")
     db_todo = models.TODO(**todo.model_dump(), owner_id=user.get('id'))
     db.add(db_todo)
     db.commit()
-    db.refresh(db_todo)
-    return db_todo
+    
 
 @router.get("/todos",status_code=status.HTTP_200_OK)
-def get_Todos(db: db_dependency):
-    result = db.query(models.TODO).all()
+def get_Todos(user:user_dependency,db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Please login first")
+    result = db.query(models.TODO).filter(models.TODO.owner_id == user.get('id')).all()
     if not result:
         raise HTTPException(status_code=404, detail="Todo not found")
     return result
 
 @router.get("/todos/{todo_id}",status_code=status.HTTP_200_OK)
-def get_Todo_By_Id(db: db_dependency, todo_id: int = Path(gt=0)):
-    result = db.query(models.TODO).filter(models.TODO.id == todo_id).first()
+def get_Todo_By_Id(user:user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Please login first")
+    result = db.query(models.TODO).filter(models.TODO.id == todo_id and models.TODO.owner_id == user.get('id')).first()
     if not result:
         raise HTTPException(404, detail='no todo with such id')
     return result
 
 @router.put("/todos/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-def update_todo(todo: TodoSchema,todo_id: int,db: db_dependency):
-    result = db.query(models.TODO).filter(models.TODO.id == todo_id).first()
+def update_todo(user:user_dependency,todo: TodoSchema,todo_id: int,db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Please login first")
+    result = db.query(models.TODO).filter(models.TODO.id == todo_id and models.TODO.owner_id == user.get('id')).first()
     if result is None:
         raise HTTPException(404,detail="todo not found")
     result.title = todo.title
@@ -74,8 +81,10 @@ def update_todo(todo: TodoSchema,todo_id: int,db: db_dependency):
     db.commit()
     
 @router.delete("/todos/{todo_id}")
-def delete_todo(db :db_dependency,todo_id: int=Path(gt=0)):
-    result = db.query(models.TODO).filter(models.TODO.id==todo_id)
+def delete_todo(user:user_dependency, db :db_dependency,todo_id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Please login first")
+    result = db.query(models.TODO).filter(models.TODO.id==todo_id and models.TODO.owner_id == user.get('id'))
     if result is None:
         raise HTTPException(404,detail="todo not found!")
     db.query(models.TODO).filter(models.TODO.id==todo_id).delete()
